+++ sqrl-testing/sqrl-integration-tests/src/test/java/com/datasqrl/graphql/MultipleSensorsTest.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2021, DataSQRL. All rights reserved. Use is subject to license terms.
+ */
+package com.datasqrl.graphql;
+
+import com.datasqrl.util.data.Sensors;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+import lombok.SneakyThrows;
+import org.junit.jupiter.api.Test;
+
+public class MultipleSensorsTest extends SensorsTest {
+
+  @SneakyThrows
+  @Test
+  public void multipleSubscribersTest() {
+    fut = execute(Sensors.INSTANCE_MUTATION);
+    events = new ArrayList<>();
+    List<CountDownLatch> latches = new ArrayList<>();
+    for (int i = 0; i < 5; i++) {
+      latches.add(subscribeToAlert(alert));
+    }
+
+    Thread.sleep(1000);
+
+    executeMutation(addReading, triggerAlertJson);
+
+    // Wait a long time for the first latch, the rest should complete soon after
+    int timeout = 120;
+    for (CountDownLatch latch : latches) {
+      latch.await(timeout, TimeUnit.SECONDS);
+      timeout = timeout / 2;
+    }
+
+    fut.cancel(true);
+
+    validateEvents();
+  }
+}
