+++ sqrl-testing/sqrl-integration-tests/src/test/resources/snapshots/com/datasqrl/graphql/MultipleSensorsTest/multipleSubscribersTest.txt
